LPS-63059 Move classes that are going to be used outside this module to the exported package
