+++ modules/util/portal-tools-data-partitioning-sql-builder-postgresql/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/postgresql/exporter/PostgreSQLDataPartitioningExporter.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.tools.data.partitioning.sql.builder.postgresql.exporter;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.internal.exporter.BaseDataPartitioningExporter;
+import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.BaseDataPartitioningExporter;
 
 import java.sql.Date;
 import java.sql.Timestamp;
