+++ modules/util/portal-tools-data-partitioning-sql-builder/src/test/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/BaseDataPartitioningExporterTestCase.java
@@ -12,10 +12,8 @@
  * details.
  */
 
-package com.liferay.portal.tools.data.partitioning.sql.builder.internal.exporter;
+package com.liferay.portal.tools.data.partitioning.sql.builder.exporter;
 
-import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.DataPartitioningExporter;
-import com.liferay.portal.tools.data.partitioning.sql.builder.exporter.DataPartitioningExporterFactory;
 import com.liferay.portal.tools.data.partitioning.sql.builder.test.util.DBProviderTestUtil;
 
 import java.sql.Connection;
